ai-bot/pyproject.toml
[project]
name = "campfire-ai-bot"
version = "0.4.0.1"
description = "AI-powered bot for Campfire chat using Claude Agent SDK with FastAPI"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "claude-agent-sdk>=0.1.4",  # Latest available version
    "anthropic>=0.39.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.0.0",  # For Operations Assistant bot
    "markitdown[pptx]>=0.0.1",  # For PowerPoint text extraction
    "pyyaml>=6.0",  # For YAML bot configuration files (v0.4.1)
    "apscheduler>=3.10.4",  # For automated reminder delivery (v0.5.1)
    "python-dateutil>=2.8.2",  # For natural language time parsing (v0.5.1)
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.3",
]
